operacion resta puesta como double

+++ Calculadora.cpp
@@ -7,7 +7,7 @@ double suma(int num1, int num2){
     return num1 + num2;
 }
 
-int resta(int num1, int num2){
+double resta(int num1, int num2){
     return num1 - num2;
 }
